requirements.txt
streamlit
langchain-community
sentence-transformers
faiss-cpu
google-generativeai